// app/lib/cashier/types.ts
export interface CashierShiftUser {
  user_id: string
  username: string
  is_primary: 0 | 1
}